+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 
 ## [1.0.17] - Unreleased
 - Update app::set_callback to reflect WidgetExt::set_callback.
+- Add app::swap_frame_type() to swap the default frame type with a new one.
+- Add app::frame_type() to get the default frame type.
+- Add enums::FrameType::swap_frames() to swap frame types.
 
 ## [1.0.16] - 2021-06-02
 - Update to latest FLTK with gdi+ support on Windows (for anti-aliased oblique lines and curves).